requirements.txt
SQLAlchemy==2.0.23
alembic==1.12.1
python-dotenv
argon2-cffi==23.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
rich==13.7.0
pytest==7.4.3
sentry-sdk==1.9.0